+++ doc/changelog.rst
@@ -17,11 +17,13 @@ API Changes
 New Features
 ************
 
-- Add :class:`~plotnine.stats.geom_qq_line` and
+- Added :class:`~plotnine.geoms.geom_qq_line` and
   :class:`~plotnine.stats.stat_qq_line`, for lines through Q-Q plots.
 
 - Added :class:`~plotnine.geoms.geom_density_2d` and
   :class:`~plotnine.geoms.geom_stat_2d`.
+
+- Added :class:`~plotnine.stats.stat_ellipse`.
 
 Bug Fixes
 *********